Merges the binding and value creation statements
This allows us to have a single statement which performs all kinds of
bindings, including partial values to partial values, expressions to
expressions, partial values and expressions on the lhs, etc.
