+++ app/oz/execution/index.js
@@ -3,7 +3,6 @@ import skip from "./skip";
 import sequence from "./sequence";
 import local from "./local";
 import binding from "./binding";
-import valueCreation from "./value_creation";
 import conditional from "./conditional";
 import patternMatching from "./pattern_matching";
 import procedureApplication from "./procedure_application";
@@ -23,7 +22,6 @@ export const executors = {
     [statementTypes.sequence]: sequence,
     [statementTypes.local]: local,
     [statementTypes.binding]: binding,
-    [statementTypes.valueCreation]: valueCreation,
     [statementTypes.conditional]: conditional,
     [statementTypes.patternMatching]: patternMatching,
     [statementTypes.procedureApplication]: procedureApplication,
