+++ app/oz/machine/statements.js
@@ -5,7 +5,6 @@ export const statementTypes = {
   sequence: "sequence",
   local: "local",
   binding: "binding",
-  valueCreation: "valueCreation",
   conditional: "conditional",
   patternMatching: "patternMatching",
   procedureApplication: "procedureApplication",
@@ -51,15 +50,6 @@ export const bindingStatement = (lhs, rhs) => {
   return Immutable.Map({
     node: "statement",
     type: statementTypes.binding,
-    lhs,
-    rhs,
-  });
-};
-
-export const valueCreationStatement = (lhs, rhs) => {
-  return Immutable.Map({
-    node: "statement",
-    type: statementTypes.valueCreation,
     lhs,
     rhs,
   });
