+++ app/oz/print/binding.js
@@ -1,10 +1,14 @@
 export default (recurse, node, identation) => {
   const ident = new Array(identation + 1).join(" ");
-  const lhs = node.getIn(["lhs", "identifier"]);
-  const rhs = node.getIn(["rhs", "identifier"]);
-  const result = `${ident}${lhs} = ${rhs}`;
+  const lhs = node.get("lhs");
+  const fullLhs = recurse(lhs, identation).full;
+  const abbreviatedLhs = recurse(lhs).abbreviated;
+  const rhs = node.get("rhs");
+  const fullRhs = recurse(rhs, identation).full;
+  const abbreviatedRhs = recurse(rhs).abbreviated;
+
   return {
-    abbreviated: result,
-    full: result,
+    abbreviated: `${ident}${abbreviatedLhs} = ${abbreviatedRhs}`,
+    full: `${ident}${fullLhs} = ${fullRhs}`,
   };
 };
