+++ specs/execution/thread_spec.js
@@ -4,6 +4,7 @@ import {
   bindingStatement,
 } from "../../app/oz/machine/statements";
 import { lexicalIdentifier } from "../../app/oz/machine/lexical";
+import { identifierExpression } from "../../app/oz/machine/expressions";
 import {
   buildState,
   buildThread,
@@ -24,7 +25,10 @@ describe("Reducing thread statements", () => {
         buildThread({
           semanticStatements: [
             buildSemanticStatement(
-              bindingStatement(lexicalIdentifier("X"), lexicalIdentifier("Y")),
+              bindingStatement(
+                identifierExpression(lexicalIdentifier("X")),
+                identifierExpression(lexicalIdentifier("Y")),
+              ),
               buildEnvironment({
                 X: buildVariable("x", 0),
                 Y: buildVariable("y", 0),
@@ -36,7 +40,10 @@ describe("Reducing thread statements", () => {
     });
     const statement = buildSemanticStatement(
       threadStatement(
-        bindingStatement(lexicalIdentifier("A"), lexicalIdentifier("B")),
+        bindingStatement(
+          identifierExpression(lexicalIdentifier("A")),
+          identifierExpression(lexicalIdentifier("B")),
+        ),
       ),
       buildEnvironment({
         A: buildVariable("a", 0),
@@ -51,8 +58,8 @@ describe("Reducing thread statements", () => {
             semanticStatements: [
               buildSemanticStatement(
                 bindingStatement(
-                  lexicalIdentifier("X"),
-                  lexicalIdentifier("Y"),
+                  identifierExpression(lexicalIdentifier("X")),
+                  identifierExpression(lexicalIdentifier("Y")),
                 ),
                 buildEnvironment({
                   X: buildVariable("x", 0),
@@ -65,8 +72,8 @@ describe("Reducing thread statements", () => {
             semanticStatements: [
               buildSemanticStatement(
                 bindingStatement(
-                  lexicalIdentifier("A"),
-                  lexicalIdentifier("B"),
+                  identifierExpression(lexicalIdentifier("A")),
+                  identifierExpression(lexicalIdentifier("B")),
                 ),
                 buildEnvironment({
                   A: buildVariable("a", 0),
