+++ specs/free_identifiers/binding_spec.js
@@ -2,19 +2,30 @@ import Immutable from "immutable";
 import { collectFreeIdentifiers } from "../../app/oz/free_identifiers";
 import { bindingStatement } from "../../app/oz/machine/statements";
 import { lexicalIdentifier } from "../../app/oz/machine/lexical";
+import {
+  literalExpression,
+  identifierExpression,
+} from "../../app/oz/machine/expressions";
+import { literalRecord } from "../../app/oz/machine/literals";
 
 describe("Collecting free identifiers in a binding statement", () => {
   beforeEach(() => {
     jasmine.addCustomEqualityTester(Immutable.is);
   });
 
-  it("collects both lhs and rhs identifiers", () => {
+  it("collects the lhs identifier and all the identifiers in the literal", () => {
     const statement = bindingStatement(
-      lexicalIdentifier("X"),
-      lexicalIdentifier("Y"),
+      identifierExpression(lexicalIdentifier("X")),
+      literalExpression(
+        literalRecord("person", {
+          age: lexicalIdentifier("A"),
+          name: lexicalIdentifier("N"),
+        }),
+      ),
     );
+
     expect(collectFreeIdentifiers(statement)).toEqual(
-      Immutable.Set(["X", "Y"]),
+      Immutable.Set(["X", "A", "N"]),
     );
   });
 });
