+++ specs/free_identifiers/exception_context_spec.js
@@ -5,6 +5,7 @@ import {
   bindingStatement,
 } from "../../app/oz/machine/statements";
 import { lexicalIdentifier } from "../../app/oz/machine/lexical";
+import { identifierExpression } from "../../app/oz/machine/expressions";
 
 describe("Collecting free identifiers in a try statement", () => {
   beforeEach(() => {
@@ -13,9 +14,15 @@ describe("Collecting free identifiers in a try statement", () => {
 
   it("collects identifiers from substatements, except for the exception identifier", () => {
     const statement = exceptionContextStatement(
-      bindingStatement(lexicalIdentifier("A"), lexicalIdentifier("B")),
+      bindingStatement(
+        identifierExpression(lexicalIdentifier("A")),
+        identifierExpression(lexicalIdentifier("B")),
+      ),
       lexicalIdentifier("X"),
-      bindingStatement(lexicalIdentifier("C"), lexicalIdentifier("X")),
+      bindingStatement(
+        identifierExpression(lexicalIdentifier("C")),
+        identifierExpression(lexicalIdentifier("X")),
+      ),
     );
 
     expect(collectFreeIdentifiers(statement)).toEqual(
@@ -25,9 +32,15 @@ describe("Collecting free identifiers in a try statement", () => {
 
   it("collects identifiers from substatements including the exception identifier if it is also used in the tried statement", () => {
     const statement = exceptionContextStatement(
-      bindingStatement(lexicalIdentifier("X"), lexicalIdentifier("B")),
+      bindingStatement(
+        identifierExpression(lexicalIdentifier("X")),
+        identifierExpression(lexicalIdentifier("B")),
+      ),
       lexicalIdentifier("X"),
-      bindingStatement(lexicalIdentifier("C"), lexicalIdentifier("X")),
+      bindingStatement(
+        identifierExpression(lexicalIdentifier("C")),
+        identifierExpression(lexicalIdentifier("X")),
+      ),
     );
 
     expect(collectFreeIdentifiers(statement)).toEqual(
