+++ specs/free_identifiers/pattern_matching_spec.js
@@ -27,8 +27,14 @@ describe("Collecting free identifiers in a pattern matching statement", () => {
         age: lexicalIdentifier("A"),
         name: lexicalIdentifier("N"),
       }),
-      bindingStatement(lexicalIdentifier("Y"), lexicalIdentifier("A")),
-      bindingStatement(lexicalIdentifier("Z"), lexicalIdentifier("N")),
+      bindingStatement(
+        identifierExpression(lexicalIdentifier("Y")),
+        identifierExpression(lexicalIdentifier("A")),
+      ),
+      bindingStatement(
+        identifierExpression(lexicalIdentifier("Z")),
+        identifierExpression(lexicalIdentifier("N")),
+      ),
     );
 
     expect(collectFreeIdentifiers(statement)).toEqual(
