--- specs/parser/statements/value_creation_spec.js
@@ -1,143 +1,0 @@
-import Immutable from "immutable";
-import { lexicalIdentifier } from "../../../app/oz/machine/lexical";
-import {
-  literalNumber,
-  literalRecord,
-  literalProcedure,
-} from "../../../app/oz/machine/literals";
-import {
-  operatorExpression,
-  literalExpression,
-} from "../../../app/oz/machine/expressions";
-import {
-  valueCreationStatementSyntax,
-  skipStatementSyntax,
-} from "../../../app/oz/machine/statementSyntax";
-import parse from "../../../app/oz/parser";
-
-describe("Parsing X=VALUE statements", () => {
-  beforeEach(() => {
-    jasmine.addCustomEqualityTester(Immutable.is);
-  });
-
-  describe("when parsing numbers", () => {
-    it("handles condensed syntax correctly", () => {
-      expect(parse("X=5")).toEqual(
-        valueCreationStatementSyntax(
-          lexicalIdentifier("X"),
-          literalExpression(literalNumber(5)),
-        ),
-      );
-    });
-
-    it("handles spaced syntax correctly", () => {
-      expect(parse("Variable = 12.0")).toEqual(
-        valueCreationStatementSyntax(
-          lexicalIdentifier("Variable"),
-          literalExpression(literalNumber(12.0)),
-        ),
-      );
-    });
-
-    it("handles quoted variable syntax correctly", () => {
-      expect(parse("`One Variable` = 152")).toEqual(
-        valueCreationStatementSyntax(
-          lexicalIdentifier("One Variable"),
-          literalExpression(literalNumber(152)),
-        ),
-      );
-    });
-  });
-
-  describe("when parsing records", () => {
-    it("handles condensed syntax correctly", () => {
-      expect(parse("X=person(name:N)")).toEqual(
-        valueCreationStatementSyntax(
-          lexicalIdentifier("X"),
-          literalExpression(
-            literalRecord("person", { name: lexicalIdentifier("N") }),
-          ),
-        ),
-      );
-    });
-
-    it("handles spaced syntax correctly", () => {
-      expect(parse("X = person(name:N)")).toEqual(
-        valueCreationStatementSyntax(
-          lexicalIdentifier("X"),
-          literalExpression(
-            literalRecord("person", { name: lexicalIdentifier("N") }),
-          ),
-        ),
-      );
-    });
-
-    it("handles quoted variable syntax correctly", () => {
-      expect(parse("`One Variable` = person(name:N)")).toEqual(
-        valueCreationStatementSyntax(
-          lexicalIdentifier("One Variable"),
-          literalExpression(
-            literalRecord("person", { name: lexicalIdentifier("N") }),
-          ),
-        ),
-      );
-    });
-  });
-
-  describe("when parsing procedures", () => {
-    it("handles condensed syntax correctly", () => {
-      expect(parse("X=proc{$ X Y} skip end")).toEqual(
-        valueCreationStatementSyntax(
-          lexicalIdentifier("X"),
-          literalExpression(
-            literalProcedure(
-              [lexicalIdentifier("X"), lexicalIdentifier("Y")],
-              skipStatementSyntax(),
-            ),
-          ),
-        ),
-      );
-    });
-
-    it("handles spaced syntax correctly", () => {
-      expect(parse("X = proc{$ X Y} skip end")).toEqual(
-        valueCreationStatementSyntax(
-          lexicalIdentifier("X"),
-          literalExpression(
-            literalProcedure(
-              [lexicalIdentifier("X"), lexicalIdentifier("Y")],
-              skipStatementSyntax(),
-            ),
-          ),
-        ),
-      );
-    });
-
-    it("handles quoted variable syntax correctly", () => {
-      expect(parse("`One Variable` = proc{$ X Y} skip end")).toEqual(
-        valueCreationStatementSyntax(
-          lexicalIdentifier("One Variable"),
-          literalExpression(
-            literalProcedure(
-              [lexicalIdentifier("X"), lexicalIdentifier("Y")],
-              skipStatementSyntax(),
-            ),
-          ),
-        ),
-      );
-    });
-  });
-
-  it("handles rhs expressions", () => {
-    expect(parse("X = 2 + 3")).toEqual(
-      valueCreationStatementSyntax(
-        lexicalIdentifier("X"),
-        operatorExpression(
-          "+",
-          literalExpression(literalNumber(2)),
-          literalExpression(literalNumber(3)),
-        ),
-      ),
-    );
-  });
-});
